+++ glean-core/ffi/glean.h
@@ -87,6 +87,64 @@ typedef const int32_t *RawIntArray;
 typedef const char *const *RawStringArray;
 
 /**
+ * ByteBuffer is a struct that represents an array of bytes to be sent over the FFI boundaries.
+ * There are several cases when you might want to use this, but the primary one for us
+ * is for returning protobuf-encoded data to Swift and Java. The type is currently rather
+ * limited (implementing almost no functionality), however in the future it may be
+ * more expanded.
+ *
+ * ## Caveats
+ *
+ * Note that the order of the fields is `len` (an i64) then `data` (a `*mut u8`), getting
+ * this wrong on the other side of the FFI will cause memory corruption and crashes.
+ * `i64` is used for the length instead of `u64` and `usize` because JNA has interop
+ * issues with both these types.
+ *
+ * ByteBuffer does not implement Drop. This is intentional. Memory passed into it will
+ * be leaked if it is not explicitly destroyed by calling [`ByteBuffer::destroy`]. This
+ * is because in the future, we may allow it's use for passing data into Rust code.
+ * ByteBuffer assuming ownership of the data would make this a problem.
+ *
+ * Note that alling `destroy` manually is not typically needed or recommended,
+ * and instead you should use [`define_bytebuffer_destructor!`].
+ *
+ * ## Layout/fields
+ *
+ * This struct's field are not `pub` (mostly so that we can soundly implement `Send`, but also so
+ * that we can verify rust users are constructing them appropriately), the fields, their types, and
+ * their order are *very much* a part of the public API of this type. Consumers on the other side
+ * of the FFI will need to know its layout.
+ *
+ * If this were a C struct, it would look like
+ *
+ * ```c,no_run
+ * struct ByteBuffer {
+ *     int64_t len;
+ *     uint8_t *data; // note: nullable
+ * };
+ * ```
+ *
+ * In rust, there are two fields, in this order: `len: i64`, and `data: *mut u8`.
+ *
+ * ### Description of fields
+ *
+ * `data` is a pointer to an array of `len` bytes. Not that data can be a null pointer and therefore
+ * should be checked.
+ *
+ * The bytes array is allocated on the heap and must be freed on it as well. Critically, if there
+ * are multiple rust packages using being used in the same application, it *must be freed on the
+ * same heap that allocated it*, or you will corrupt both heaps.
+ *
+ * Typically, this object is managed on the other side of the FFI (on the "FFI consumer"), which
+ * means you must expose a function to release the resources of `data` which can be done easily
+ * using the [`define_bytebuffer_destructor!`] macro provided by this crate.
+ */
+typedef struct {
+  int64_t len;
+  uint8_t *data;
+} ByteBuffer;
+
+/**
  * A FFI-compatible representation for the PingUploadTask.
  *
  * This is exposed as a C-compatible tagged union, like this:
@@ -143,7 +201,7 @@ typedef struct {
   FfiPingUploadTask_Tag tag;
   char *document_id;
   char *path;
-  char *body;
+  ByteBuffer body;
   char *headers;
 } FfiPingUploadTask_Upload_Body;
 
@@ -278,7 +336,7 @@ char *glean_experiment_test_get_data(FfiStr experiment_id);
 
 uint8_t glean_experiment_test_is_active(FfiStr experiment_id);
 
-FfiPingUploadTask glean_get_upload_task(void);
+void glean_get_upload_task(FfiPingUploadTask *result);
 
 /**
  * # Safety
@@ -475,7 +533,19 @@ uint8_t glean_on_ready_to_submit_pings(void);
 
 char *glean_ping_collect(uint64_t ping_type_handle, FfiStr reason);
 
-void glean_process_ping_upload_response(FfiPingUploadTask task, uint32_t status);
+/**
+ * Process and free a `FfiPingUploadTask`.
+ *
+ * We need to pass the whole task instead of only the document id,
+ * so that we can free the strings properly on Drop.
+ *
+ * After return the `task` should not be used further by the caller.
+ *
+ * # Safety
+ *
+ * A valid and non-null upload task object is required for this function.
+ */
+void glean_process_ping_upload_response(FfiPingUploadTask *task, uint32_t status);
 
 void glean_quantity_set(uint64_t metric_id, int64_t value);