+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean/compare/v30.0.0...master)
+[Full changelog](https://github.com/mozilla/glean/compare/v30.1.0...master)
+
+# v30.1.0 (2020-05-22)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v30.0.0...v30.1.0)
+
+* Android & iOS
+  * Ping payloads are now compressed using gzip.
+* iOS
+  * `Glean.initialize` is now a no-op if called from an embedded extension. This means that Glean will only run in the base application process in order to prevent extensions from behaving like separate applications with different client ids from the base application. Applications are responsible for ensuring that extension metrics are only collected within the base application.
+* Python
+  * `lifetime: application` metrics are now cleared after the Glean-owned pings are sent,
+    after the product starts.
+  * Glean Python bindings now build in a native Windows environment.
+  * BUGFIX: `MemoryDistributionMetric` now parses correctly in `metrics.yaml` files.
+  * BUGFIX: Glean will no longer crash if run as part of another library's coverage testing.
 
 # v30.0.0 (2020-05-13)
 
@@ -17,6 +32,13 @@
   * Refactor the ping uploader to use the new upload mechanism.
 * iOS:
   * Refactor the ping uploader to use the new upload mechanism.
+
+# v29.1.1 (2020-05-22)
+
+[Full changelog](https://github.com/mozilla/glean/compare/v29.1.0...v29.1.1)
+
+* Android
+  * BUGFIX: Fix a race condition that leads to a `ConcurrentModificationException`. [Bug 1635865](https://bugzilla.mozilla.org/1635865)
 
 # v29.1.0 (2020-05-11)
 
@@ -109,7 +131,7 @@
     were unintentionally public, have been made private.
   * Most Glean work and I/O is now done on its own worker thread. This brings the parallelism Python in line with the other platforms.
   * The timing distribution, memory distribution, string list, labeled boolean and labeled string metric types are now supported in Python ([#762](https://github.com/mozilla/glean/pull/762), [#763](https://github.com/mozilla/glean/pull/763), [#765](https://github.com/mozilla/glean/pull/765), [#766](https://github.com/mozilla/glean/pull/766))
-  
+
 # v25.1.0 (2020-02-26)
 
 [Full changelog](https://github.com/mozilla/glean/compare/v25.0.0...v25.1.0)